fix: поправил классы UserService/UserStorage
