+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,8 +2,8 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.dao.FilmStorage;
-import ru.yandex.practicum.filmorate.dao.UserStorage;
+import ru.yandex.practicum.filmorate.dao.InMemoryFilmStorage;
+import ru.yandex.practicum.filmorate.dao.InMemoryUserStorage;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
@@ -11,6 +11,7 @@ import java.util.*;
 
 import java.util.ArrayList;
 import java.util.Comparator;
+import java.util.stream.Collectors;
 
 import static java.util.Collections.sort;
 
@@ -18,79 +19,68 @@ import static java.util.Collections.sort;
 public class FilmService {
 
     @Autowired
-    FilmStorage filmStorage;
+    private InMemoryFilmStorage inMemoryFilmStorage;
 
     @Autowired
-    UserStorage userStorage;
+    private InMemoryUserStorage inMemoryUserStorage;
 
     public Film get(long filmId) {
-        final Film film = filmStorage.get(filmId);
+        final Film film = inMemoryFilmStorage.get(filmId);
         if (film == null) {
             throw new NotFoundException("Film with id=" + filmId + " not found");
         }
-        return filmStorage.get(filmId);
+        return inMemoryFilmStorage.get(filmId);
     }
 
     public Film save(Film film) {
-        return filmStorage.saveFilm(film);
+        return inMemoryFilmStorage.saveFilm(film);
     }
 
     public ArrayList<Film> getFilms() {
-        return filmStorage.getFilms();
+        return inMemoryFilmStorage.getFilms();
     }
 
     public void update(Film film) {
-        filmStorage.update(film);
+        inMemoryFilmStorage.update(film);
     }
 
     public void addLike(long filmId, long userId) {
-        if (userStorage.get(userId) == null) {
+        if (inMemoryUserStorage.get(userId) == null) {
             throw new NotFoundException("User with id=" + userId + " not found");
-        } else if (filmStorage.get(filmId) == null) {
+        } else if (inMemoryFilmStorage.get(filmId) == null) {
             throw new NotFoundException("Film with id=" + filmId + " not found");
         } else {
-            Film film = filmStorage.get(filmId);
-            User user = userStorage.get(userId);
-            filmStorage.addLike(film, user);
+            Film film = inMemoryFilmStorage.get(filmId);
+            User user = inMemoryUserStorage.get(userId);
+            inMemoryFilmStorage.addLike(film, user);
         }
     }
 
     public void deleteLike(long filmId, long userId) {
-        if (userStorage.get(userId) == null) {
+        if (inMemoryUserStorage.get(userId) == null) {
             throw new NotFoundException("User with id=" + userId + " not found");
-        } else if (filmStorage.get(filmId) == null) {
+        } else if (inMemoryFilmStorage.get(filmId) == null) {
             throw new NotFoundException("Film with id=" + filmId + " not found");
         } else {
-            User user = userStorage.get(userId);
-            Film film = filmStorage.get(filmId);
-            filmStorage.deleteLike(film, user);
+            User user = inMemoryUserStorage.get(userId);
+            Film film = inMemoryFilmStorage.get(filmId);
+            inMemoryFilmStorage.deleteLike(film, user);
         }
     }
 
+    Comparator<Film> comparator = new Comparator<Film>() {
+        @Override
+        public int compare(Film o1, Film o2) {
+            return o2.getUserIds().size() - o1.getUserIds().size();
+        }
+    };
+
     public List<Film> getFilmLikes(int count) {
-        int countFilms = 9;
-        if (count > 0) {
-            countFilms = count - 1;
-        }
-        List<Film> films = filmStorage.getFilms();
-        List<Film> sortedFilms = new ArrayList<>();
-        Collections.sort(films, new Comparator<Film>() {
-            @Override
-            public int compare(Film o1, Film o2) {
-                return o2.getUserIds().size() - o1.getUserIds().size();
-            }
-        });
-        if (countFilms > films.size()) {
-            countFilms = films.size() - 1;
-        }
-        if (countFilms == 0) {
-            sortedFilms.add(films.get(0));
-        } else {
-            for (int i = 0; i <= countFilms; i++) {
-                sortedFilms.add(films.get(i));
-            }
-        }
-        return sortedFilms;
+        List<Film> films = inMemoryFilmStorage.getFilms();
+        Collections.sort(films, comparator);
+        return films.stream()
+                .limit(count)
+                .collect(Collectors.toList());
     }
 
 }
